+++ libs/ngx-editorjs-plugins/src/lib/services/plugins.module.ts
@@ -1,6 +1,9 @@
 import { NgModule } from '@angular/core';
 import { NgxEditorJSPluginService } from './plugins.service';
 
+/**
+ * This module provides the plugin service used by NgxEditorJS.
+ */
 @NgModule({
   providers: [NgxEditorJSPluginService]
 })