+++ libs/ngx-editorjs-plugins/src/lib/services/plugins.service.ts
@@ -29,28 +29,28 @@ export class NgxEditorJSPluginService {
 
   /**
    * Add a plugin to the store
-   * @param name
-   * @param tool
+   * @param key The key for the map to store the plugin
+   * @param plugin The plugin instance to add to the service
    */
-  public add(name: string, tool: BasePlugin) {
-    this.pluginsMap[name] = tool;
+  public add(key: string, plugin: BasePlugin) {
+    this.pluginsMap[key] = plugin;
   }
 
   /**
    * Remove a plugin from the store
-   * @param name
+   * @param key
    */
-  public remove(name: string) {
-    this.pluginsMap[name] = null;
-    delete this.pluginsMap[name];
+  public remove(key: string) {
+    this.pluginsMap[key] = null;
+    delete this.pluginsMap[key];
   }
 
   /**
    * Get a single plugin from the map
    * @param name
    */
-  public get(name: string) {
-    this.pluginsMap[name];
+  public get(key: string): BasePlugin {
+    return this.pluginsMap[key];
   }
 
   /**
@@ -70,18 +70,17 @@ export class NgxEditorJSPluginService {
         return !exclude.includes(key);
       })
       .reduce((finalTools, [key, plugin]) => {
-        let p;
         if (plugin.shortcut) {
-          p = {
+          return {
             [key]: {
               class: plugin.plugin(),
               shortcut: plugin.shortcut()
-            }
+            },
+            ...finalTools
           };
         } else {
-          p = { [key]: plugin.plugin() };
+          return { [key]: plugin.plugin(), ...finalTools };
         }
-        return { ...finalTools, ...p };
       }, {});
   }
 }